Move site launch gating hook to client/lib/
Rename to use-site-launch-gating-variant.ts and move from
client/lib/explat/ to client/lib/. Add a temporary exception in the
Dashboard ESLint config for the import.
